Subtle local history indicator and fix unsend image
